Web front end, caching and efficiency
The api has been changed to use directorys params instead of post querys
and the domain has been set to //.uk.to/api/ instead of its own sub
domain as it was the simple solution to CORs errors for XSS for
requesting JSON across domains. The web front end for search is
functional just not pretty and SASS is awesome!